// src/app/shared/model/DocChecklist/RFADocChecklist.Model.ts
export class RFADocChecklist {
    TaskId: any;
    InstanceId: any;
    TaskListId: any;
    TransactionNo: any;
    Reason: any;
    WFCode: any;
    ListValue: any;
    ApprovedBy: any;
    Notes: any;
    RowVersion: any;
    RequestRFAObj: any;
    constructor() {
        this.TaskId = "",
            this.RowVersion = ""
    }
}  